Rename folders and files for clarity
