+++ features/Resume/Content.tsx
@@ -8,7 +8,7 @@ import { Stepper, StepperItem, StepperStep } from '@/components/Stepper'
 import { Text } from '@/components/Text'
 import biocodeLogomark from '@/public/biocode/hero/joonassandell-biocode-logomark.png'
 import mediasignalLogomark from '@/public/mediasignal/joonassandell-mediasignal-logomark.png'
-import profile from '@/public/images/joonassandell-2.jpg'
+import profile from '@/public/common/joonassandell/joonassandell-portrait-nature.jpg'
 
 export const ResumeContent = () => (
   <div className="Template-content grid-col grid-col:11@s grid-col:9@m">
