+++ public/common/decoration/line-ball.png
