+++ public/common/decoration/line-cube.png
