+++ stylesheets/index.scss
@@ -4,9 +4,9 @@
 
 @layer base {
   @include use('tokens');
-  @include use('generic/base');
-  @include use('generic/fonts');
-  @include use('generic/typography');
+  @include use('common/base');
+  @include use('common/fonts');
+  @include use('common/typography');
 }
 
 @layer vendors {
